requirements.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
attrs==25.4.0
audioread==3.1.0
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
datasets==4.6.1
decorator==5.2.1
dill==0.4.0
einops==0.8.2
einx==0.4.1
ema-pytorch==0.7.9
encodec==0.1.1
filelock==3.20.0
Flask==3.1.3
flatbuffers==25.12.19
fonttools==4.61.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
h11==0.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.34.4
hydra-core==1.3.2
idna==3.11
itsdangerous==2.2.0
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
loguru==0.7.3
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.8
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
multiprocess==0.70.18
networkx==3.6.1
numba==0.64.0
numpy==2.4.2
omegaconf==2.3.0
onnxruntime==1.23.2
packaging==26.0
pandas==3.0.1
pillow==12.0.0
platformdirs==4.9.2
pooch==1.9.0
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
pypinyin==0.55.0
python-crfsuite==0.9.12
python-dateutil==2.9.0.post0
PyYAML==6.0.3
razdel==0.5.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
ruaccent==1.5.8.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentencepiece==0.2.1
sentry-sdk==2.54.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sounddevice==0.5.5
soundfile==0.13.1
soxr==1.0.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torchdiffeq==0.2.5
tqdm==4.67.3
transformers==4.57.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
vocos==0.1.0
wandb==0.25.0
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.6
win32_setctime==1.2.0
x-transformers==2.16.2
xxhash==3.6.0
yarl==1.23.0
qwen-tts